debugging for non-interactive usage of pip
