+++ pip_check_reqs/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '2.0.2'
+__version__ = '2.0.3'
